requirements.txt
# SCRAM Data Collection Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data storage
pyarrow>=13.0.0  # For Parquet format
fastparquet>=2023.7.0  # Alternative Parquet engine

# Optional: Database support (for future modules)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Jupyter for exploration (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Future modules dependencies (commented for now)
# Feature Engineering
# scikit-learn>=1.3.0
# scipy>=1.11.0

# Model Training
# xgboost>=2.0.0
# lightgbm>=4.0.0
# tensorflow>=2.13.0  # For LSTM
# torch>=2.0.0  # Alternative for LSTM

# Monitoring and logging
# wandb>=0.15.0
# tensorboard>=2.13.0
